Add IBM-3 alignment model

+++ src/sw_models/aSourceHashF.h
@@ -18,11 +18,7 @@ along with this program; If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
-//--------------- Include files --------------------------------------
-
 #include "sw_models/aSource.h"
-
-//--------------- Classes ---------------------------------------------
 
 class aSourceHashF
 {
@@ -32,12 +28,12 @@ public:
     bucket_size = 1
   };
 
-  size_t operator()(const aSource& a1) const
+  size_t operator() (const aSource& a1) const
   {
-    return (size_t)(16384 * a1.j) + (256 * a1.slen) + a1.tlen;
+    return (size_t)(16384 * a1.j) + ((size_t)256 * a1.slen) + a1.tlen;
   }
 
-  bool operator()(const aSource& left, const aSource& right) const
+  bool operator() (const aSource& left, const aSource& right) const
   {
     return left < right;
   }
